Add diff.EditList to provide for a list of Edit instances
Often a file difference involves more than one modified region, in
which case we need more than one Edit instance to describe the full
change that was made.

List<Edit> could be used with any generic List implementation, but
we may want to offer additional API functionality that is Edit list
specific, so I'm creating a specialized List implementation for it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>