Add diff.RawText to index a file content for later compares
The RawText class converts a byte[] into an indexed map of file
lines, assuming the source content is a UNIX formatted text file.
This fits in line with Git's usage of libxdiff, where typically
the file content is treated as a UNIX formatted text file, unless
binary content was previously detected in the file stream.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>