Teach FileHeader, HunkHeader how to create an EditList
The EditList type along with its member Edit instances can be very
useful when mining a patch for information about the change it is
trying to represent to the content.  This can be useful to create a
modified version of the patch, such as with larger or smaller number
of context lines.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>