FindBugs: Disable clone() idiom tests
We tend not to implement Cloneable, but instead define our own
form of clone() on objects that we want to permit making clones
of.  This is because Cloneable is a rather poor interface in
the JVM with its implicit field copying, but the method name
is the right name for us.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>