Don't log TransportGitSsh error stream to JVM stderr
This is an API function, not a user interface routine.  We should not
be printing messages from the remote side to our console.  Instead we
should buffer them, and allow a TransportException to be thrown which
contains the message data.  Higher level user interface routines
can then catch that exception, and display the message, if any, in
a form that is appropriate to the context.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>