Add ref rename support
Now refs can be renamed. The intent is that rename should be
safe. Only the named refs and associated logs are updated. Any
symbolic refs referring to the renames branches are unaffected,
except HEAD, which is updated if the branch it currently refers to
is being renamed.

In order to perform operations like rename from a to a/b (or the
reverse of a/b to a), we perform the rename similar to the way C
Git does it.  One update at a time using temporary intermediate
file names.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>