Use a UTC relative time zone for PersonIdent
The timezone used in Git objects is not the user's preferred
timezone, but rather a numeric form stating the offset relative to
UTC in hours and minutes.  Any special rules like daylight savings
is already accounted for and cannot be deduced from the time zone
portion of the git timestamp.

As an example, a committer in Sweden will have +0100 as his/her
timezone in winter and +0200 in summer.

We hereby abandon the attempts to guess a real timezone from the
UTC offset.  When creating a person ident a real timezone can be
used, but it's identity will be lost when externalizing and only
the UTC offset will be left.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>