Add timeouts to anonymous git:// daemon
If the initial command line isn't received within the configured
timeout period, the connection is dropped, and the service thread
is able to terminate.

The timeout is also pushed down to the service implementations,
so they can abort if the client doesn't speak to them within the
configured time span.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>