Add support for remote.name.pushurl
In C Git commit 203462347fce Michael J Gruber added support for a
new URL key within a remote block, permitting a different URL to
be used for push than for fetch.  In the commit message he cites
an example where fetch runs over git://, but push uses ssh://,
as the git:// protocol has lower connection setup overheads.

This change complicates the Transport API as now we must know
in advance when the Transport.open() call is made what type of
operation the caller wants to perform, so we know which config
key to honor when constructing the Transport objects.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>