Update Index (refresh) doesn't do anything obvious
The bug reported against EGit is really a JGit bug, and is quite
easy to fix, which is why I do it rather than drop the GitIndex
usage and replace with DirCache.

Update of the sha1 variable needs to be delayed until after the
comparsion is done against the old value, otherwise modified is not
set and the operation does not complete as expected if the SHA-1
value has changed.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>