Fixed the test testUpdateRefLockFailureLocked broken on the Windows platform
On Windows the lock is aquired in the test testUpdateRefLockFailureLocked
(RefUpdateTest) for the file but it never released. This causes test
failure during resource cleanup, since locked file could not be deleted
on Windows. This commits adds unlock operation in the finally block,
so there is no more cleanup problem.

Signed-off-by: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>