Add getEncoding() to RevCommit to discover the encoding
If an application needs to parse the raw buffer by hand it might
benefit from knowing the encoding of the commit.  We can make it
available to them through a getEncoding() method, using the same
logic we already use for getFullMessage() and getShortMessage(),
but this is still only an estimate based on the "encoding" header
and may not be reality if the commit is horribly malformed.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>