Support extracting emails from Signed-off-by lines
A typical Signed-off-by footer line can be seen below in this
message, and it clearly contains an email address (or at least
a user@domain sort of reference to the identity).  This identity
may need to be parsed out of a footer line by application code,
to match the Signed-off-by line back to the author or the committer
of the commit, or to some external datastore.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>