Add BlobBasedConfig to read configuration stored in blobs
The class BlobBasedConfig allows loading files stored in blobs inside
the repository, rather than from a working copy. This class can be
used to access the .gitmodules file stored inside a bare repository.

To simplify the implementation of this class getRepository() is
added to the legacy Commit type, and the type signature of the
last argument in TreeWalk.forPath(db, path, trees) is changed from
AnyObjectId[] to AnyObjectId..., allowing vararg style calling.

Signed-off-by: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>