Support for delegating tree iterators
This patch introduce some methods that simplify creation of tree
iterators that wraps other iterators and which are located in
other package. Such iterators need to efficiently access the
name component of the path of wrapped iterator and wrapped bits.
The patch also adds a method that ensuring that path buffer has
a requried capacity when the required capacity is known in advance.

Signed-off-by: Constantine Plotnikov <constantine.plotnikov@gmail.com>
[sp: fixed minor style nits]
[sp: called new method getEntryRawMode, not getEntryFileModeBits]
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>