Avoid growing WorkingTreeIterator's path multiple times
If the current path buffer isn't big enough for the current file
name to be appended onto it, we can immediately grow it to the
correct target size by using the new ensurePathCapacity method.
This may save us from growing the buffer twice in rapid succession
if the path name component is really long, and the current buffer
is still the fairly small default size.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
CC: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>