Allow http_proxy values without protocol-part
jgit reads the environment variable http_proxy to configure the
JRE's URL library for HTTP traffic through the proxy in the same
way as other applications that are using the libcurl library.
libcurl assumes "http://" when the protocol is not specified in
http_proxy, so jgit should do the same.

Thus http_proxy=proxy:8080 is now valid.

Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>