Make Daemon's exportAll check not require synchronization
A volatile boolean can be read more efficiently than using a
synchronized monitor on "this", as it only requires that the
load of the boolean be examining the most recent copy from the
memory bus.  Since the value of the boolean has no further
impact on the object state, we don't need to worry about more
than ensuring we see the current value of it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>