Refactor ref advertisement code from server implementations
The RefAdvertiser implements the common logic from both UploadPack
and ReceivePack, as the code here was really very similar.  By using
a support class we can also cleanup the nasty passing of parameters
for temporary data, and instead rely on the instance members of the
support class to handle it for us.

A minor change in functionality exists in this change, we now make
a temporary RevFlag and tag all advertised objects in ReceivePack.
This is a pointless operation that currently wastes a very small
amount of CPU time during setup of the advertisement, but is so
minor that its just not worth optimizing out.  The code reuse is
much more important for maintenance.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>