Don't advertise HEAD from ReceivePack
The HEAD ref cannot be pushed to by a client, if it is a detached HEAD
the client shouldn't be permitted to change it, if it is a symref to
another ref then the client should update the destination ref and not
the symref.  Instead offer the HEAD ref as a ".have" line, which is an
invalid ref that the client can't update but still lets the client know
we have its target object reachable.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
[rr. Use equals for comparing ref name and origName]
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>