Use cached Repository instances when resolving alternates
Using a cached Repository instance when opening an alternate object
database ensures that if multiple repositories share the same base
repository (e.g. 4 forks of the Linux kernel all share the same
base "linus-2.6.git" repository) then the base object directory is
only monitored and tracked once in the JVM, instead of 4 times.  In
the case of the WindowCache, this means we will have greater reuse
of the windows from the base, as they won't be loaded 4 different
times under different keys.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>