Send .have lines in ReceivePack for alternate repositories
If we are borrowing objects from an alternate repository we should
send the alternates' refs, but under the magical ".have" name, so
that a push client won't transmit objects that our alternate already
contains for us.  To save bandwidth during the advertisement phase
of the connection any duplicates are skipped.  This way we don't
send a ".have" if we have already sent that same object name under
a different ref, or under a prior have line.

In 5d3a6706eef2 we introduced the client side support for this,
and now we introduce the corresponding server side support.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>