Use NB.readFully(File) to slurp complete file contents
When we want the entire file contents and its likely a very small
file (e.g. a ref, configuration file) NB.readFully(File) is an
easier method to get the content then using FileInputStream from
application code.

When looking at a file that we believe should contain only a SHA-1
or be a symbolic ref to another file within this repository,
we use a maximum read cap of 4096 bytes for the slurp, so that
we don't accidentally allocate a very large buffer and blow the
JVM heap up.  This reasonableness cap is particularly important
for the RepositoryCache.FileKey.isGitRepository() test logic, in
case it comes across a file called HEAD that isn't really part of
a Git repository.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>