Clarify section and subsection values in Config code
When I first wrote this code I wasn't clear in my intent; the section
name was sometimes called "base" and the subsection name was somtimes
called "extendedBase".  This is nuts, everywhere else in our code we
use "section" and "subsection".  Renaming the fields makes the code
much easier to follow.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
[rr. resolved trivial conflict after 1f899f9 by Constantine Plotnikov.]
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>