Cleanup Config's MAGIC_EMPTY_VALUE to be more safe
The magic value "%%magic%%empty%%" is just too magic; if it ever
did appear as a value in a key string Config would have treated
it as a true value instead of as a string value.  We also had to
special case conversions of it to the empty string in a string
context.  Instead we create a special String object and use reference
equality against that to indicate the magic empty value.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
[rr. use new String() instead of StringBuilder and silence FB on it].
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>