Do not lowercase subsection names when accessing them in config files.
The entries are stored in Config.byName without coversion of
the subsection name to the lowercase. But the methods
setStringList(...) and getRawEntry(...) were converting subsection
names to the lowercase, thus making it impossible to access
values in subsection with names that contained upppercase characters.
This patch removes conversion to the lowercase and introduce
the methods that appropriately concatentates the key. This key
is now used for all map accesses.

Signed-off-by: Constantine Plotnikov <constantine.plotnikov@gmail.com>
[rr. Shortened first line]
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>