Refactor Config hierarchy to make IO more explicit
Accessing a config file may or may not require IO, e.g. if we are
reading a file that was already loaded from a blob in the repository
then there is no additional IO required.  Moving the IO portions out
of the base Config class helps to isolate it to only where we really
know we have to do IO, making it easier to handle the IO conditions.

This makes it easier to deal with test cases for config files, as
we no longer need to perform local disk IO just to test the parse
and formatting logic.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
CC: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>