Remove the map lookup for values in Config
Update operations are already O(N) time in the config file, as we
scan through every entry in the file, in order, looking to see if
it matches the key we need to mutate.  As most configurations are
rather small in size the running time is trivial.  Doing the same
for lookup simplifies our code considerably.

Applications really want a different type of fast lookup strategy.
Instead of by key lookups they want something like CoreConfig
or RemoteConfig to be cached and quickly obtainable on demand,
where the all relevant values have been parsed from their string
content representation into the application specific data types.
This change does not provide that functionality, but rather assumes
we will do so in the near future.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>