Make Config thread safe by using copy-on-write semantics
It is relatively rare that the Config instance is modified from
application code.  When it is, we usually are working on a small
set of values read from .git/config or ~/.gitconfig.  Making the
class thread-safe through basic copy-on-write semantics for its
inner collection of entries simplifies the locking and permits us
to optimize for the much more common read path.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>