Support cached application models in a Config
The Config.SectionParser acts as a key into a cache of model object
instances, but also knows how to create its model instance if the
cache does not contain the corresponding instance.

Because we don't keep track of which keys were accessed by a parser
when it created its model, we discard the entire cache anytime any
key is modified, forcing all models to reparse on the next access.

Both CoreConfig and TransferConfig are trivial to switch to this
method of access, although the typical get path is quite a bit longer
as we now need to do a hash lookup instead of a getfield instruction.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
CC: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>