Add no fallback constructor to FileBasedConfig
Sometimes when reading a FileBasedConfig the caller does not want any
fallback lookup.  Instead always passing a null first parameter to the
constructor, we can use a 1-arg constructor that takes only the File
path for this file.  This simplifies a number of cases within the JGit
library and test cases, as well as some application code.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>